Move SEAL BFV encryptions to own test file
