+++ logproof/src/lib.rs
@@ -48,7 +48,7 @@ pub mod linear_algebra;
  */
 mod linear_relation;
 pub use linear_relation::{
-    LogProof, ProverKnowledge as LogProofProverKnowledge,
+    Bounds, LogProof, ProverKnowledge as LogProofProverKnowledge,
     VerifierKnowledge as LogProofVerifierKnowledge,
 };
 
@@ -62,3 +62,5 @@ pub mod fields;
  */
 pub mod math;
 mod transcript;
+
+pub use transcript::LogProofTranscript;
